feat: move to transform.shrink
