+++ dictknife/transform.py
@@ -62,3 +62,30 @@ def only_num(d):
 
 def only_str(d):
     return {k: v for k, v in d.items() if isinstance(v, str)}
+
+
+def shrink(d, *, max_of_string_length=100, cont_suffix="...", max_of_list_length=3):
+    # todo: random select
+    # todo: cont suffix for list
+
+    from dictknife.accessing import ImmutableModifier
+
+    modifier = ImmutableModifier()
+
+    def _map(d):
+        if isinstance(d, (list, tuple)):
+            xs = d
+            if len(xs) > max_of_list_length:
+                xs = xs[:max_of_list_length]
+            return modifier.modify_list(_map, xs)
+        elif hasattr(d, "keys"):
+            return modifier.modify_dict(_map, d)
+        elif isinstance(d, str):
+            s = d
+            if len(s) > max_of_string_length:
+                s = s[:max_of_string_length] + cont_suffix
+            return s
+        else:
+            return d
+
+    return _map(d)
